+++ src/pseudoTerminal.ts
@@ -138,7 +138,7 @@ export async function replaceVariablesRecursive<T>(object: T, handler: (value: s
 
 export async function createTerminal(options: TerminalOptions): Promise<SSHPseudoTerminal> {
     const { connection } = options;
-    const { actualConfig, client } = connection;
+    const { actualConfig, client, shellConfig } = connection;
     const onDidWrite = new vscode.EventEmitter<string>();
     const onDidClose = new vscode.EventEmitter<number>();
     const onDidOpen = new vscode.EventEmitter<void>();
@@ -169,17 +169,12 @@ export async function createTerminal(options: TerminalOptions): Promise<SSHPseud
                 let SHELL = '$SHELL';
                 // Add exports for environment variables if needed
                 const env = mergeEnvironment(connection.environment, options.environment);
-                commands.push(environmentToExportString(env));
+                commands.push(environmentToExportString(env, shellConfig.setEnv));
                 // Beta feature to add a "code <file>" command in terminals to open the file locally
                 if (getFlagBoolean('REMOTE_COMMANDS', false, actualConfig.flags)[0]) {
                     const profilePathEnv = env.find(e => e.key === 'KELVIN_SSHFS_PROFILE_PATH');
                     if (!profilePathEnv) throw new Error(`Missing KELVIN_SSHFS_PROFILE_PATH environment variable`);
-                    // For bash
-                    commands.push(`export ORIG_PROMPT_COMMAND="$PROMPT_COMMAND"`);
-                    commands.push(`export PROMPT_COMMAND='source "${profilePathEnv.value}" PC; $ORIG_PROMPT_COMMAND'`);
-                    // For sh
-                    commands.push(`export OLD_ENV="$ENV"`); // not actually used (yet?)
-                    commands.push(`export ENV="${profilePathEnv.value}"`);
+                    commands.push(shellConfig.setupRemoteCommands(profilePathEnv.value));
                 }
                 // Push the actual command or (default) shell command with replaced variables
                 if (options.command) {